refactor: create view package
