+++ src/main/java/race/view/outbound/OutputView.java
@@ -1,4 +1,4 @@
-package race.outbound;
+package race.view.outbound;
 
 import race.Car;
